Remove cleaning all autoscale launch configs
This is more of a pain than a benefit. It tries to delete all ASG LCs
blindly, so most of the time it spends its time silently failing because
the launch configs are in use. It would be nice to make this a batch
operation but that's not possible in the boto API, nor does it seem
particularly easy to find unused launch configs in the boto API.
Additionally, since this was introduced we now properly delete our
launch configs when we're done using them, so it should be less of a
problem now. Even more annoying, this also introduces a race condition
where we might delete a launch config after its been created but before
its been added to an autoscaling group.

In the event that we actually do need to cleanup launch configs, it's
possible to do that via disco_autoscale.py's command line interface.