Moved more configuration magic into disco_config package.
Removed duplicative environment-specific option-finding code out
of disco_ssm and disco_elasticache, and put it into a new subclass
of ConfigParser, which disco_config.read_config will now return in
place of ConfigParser.

In parallel, updated test-patching code to use a subclass of the new
ConfigParser subclass, so that tests and live code are reading off
the same hymnal.

For sneaky reasons related to why this seemed like a good idea in the
first place, also added a new method for templating s3 bucket names.