Refactored data pipeline modification to allow passing in subnet.
Eliminated the shim add_log_location_param method, converted its tests
to tests for the new method, and covered new cases.

Added subnet_id arguments to places where setting the content of a pipeline
is allowed, and refactored a bit under the hood to make sure that the update
is made in all the places where you would expect it to be.

Incidentally fixed an arguable bug with the behavior of update_content.