Allow Excluding AMIs From Cleanup
Added a new cli option to `disco_bake.py cleanupamis` named
`--exclude-amis`. This option takes a comma seperated list of AMIs that
should be excluded from deletion by the `cleanupamis` command. This is
needed in order to support blacklisting of AMIs that should not be
deleted because they are in use in production.