Handle Bad SSM S3 Bucket Configuration
If the S3 bucket for storing SSM commands is not correctly configured,
we should fallback to using the limited storage that is available in
SSM. This might be a problem if for example the S3 bucket does not exist
or we do not have permission to access it.

Also refactored DiscoSSM to access S3 via Boto3 instead of using
DiscoCreds.