Fix disco_iam.py List Roles Bug
The default max items is 100. Let's bump that up to 500 so that this
works again. In the future, we should either write some function that
automatically paginates through the responses for boto, or we should
fast-track the effort to rewrite this module with boto3 and reuse the
existing max items function for that.