AL-1031 Remove Classic Deployment
Removed all classic deployment code. Added exception that is thrown if
an unknown deployment strategy is used (which is any that isn't
blue_green). Also removed tests for the classic deployment code and
added two tests to verify that the exception is thrown. Also updated the
README to no longer discuss classic deployment.