AL-1220 Fix Integration Tests
There were a few things breaking the integration tests:

* A bunch of tests were using mhcntp, which no longer exists. Changed
that to mhcbanana.
* The tests were expecting mhcadminproxy to come up, but the pipeline
they were using didnt have mhcadminproxy. Added mhcadminproxy to the
pipeline.
* The integration tests were calling `disco_aws.py isready`, but not
handling when that command would return an error code. Changed how we
were calling that method to not error out if the command failed.

Additionally, enabled the identity provider integration test, now that
we use them, and fixed a typo.