AL-1230 Parallelize the Integration Tests
Changed the integration tests to be able to run in parallel. For the
most part, this only required marking the fixtures of the individual
modules as shared so that running the tests in parallel would not create
multiple copies of the same fixtures.

However, this did require some significant changes to both the
disco_accounts.py script as well as its tests. The tests were rewritten
so that each test would use a different test account. This was
accomplished by not interpolating the test account name into the tests
until the runtime of the tests, and then setting a unique test account
name for each test.

Additionally, the disco_accounts.py code was modified such that it would
not attempt to refresh all existing groups and accounts everytime an
account was edited. This functionality was added to prevent an
account/group from being transfered to Glacier by continually touching
them all. This isn't actually necessary, as one could, and should,
disable transfering keys in their credentials buckets to Glacier. We've
done this internally already.

Finally, `rake test:integration` now sets the parallelization flags so
that nosetests runs in parallel. One important note is the process
timeout flag. That flag is such a high value so that our longer tests
(bake/vpc/provision/spinup) has a chance to complete.