Support Splitlines in disco_vpc.ini
It's legitimate to split an entry in an `.ini` file over multiple lines.
We'd really like to do this for some of the configuration in
disco_vpc.ini, so let's support that in our parsing of that file.

Now we'll split on all whitespace for the gateway options, which will
remove both the spaces as well as the newline characters.