fix issue where instance types change m3.large when updating existing group
Asiaq was trying to read the existing instance type of a group from its launch config
but Elastigroups don't have launch configs so it was failing and returning the default.
This caused a number of problems so implement the get_launch_config method of Elastigroups
to return the instance type for the rest of Asiaq to use.

Also add back support for modifying instance type of existing groups that was disabled previously
which tracking down this bug