Fixing Incorrect Code Coverage Statistics
Apparently there is a race condition where if nosetests is configured to
delete coverage files, it will delete some coverage files before the
code coverage statistics are calculated. This only appears to happen if
the tests are run in parallel. Removing the `code-erase` option fixes
this issue, and we already ignore the coverage files that are generated,
so this seems like a good compromise.