Fix: Error On AMI Not Found
Actually throw an exception if the requested AMI isn't found because
that's a real error.

I was considering upgrading the other case into throwing an exception as
well, but I think that might break a lot of Jenkins jobs that expect
that error message to not actually return a bad exit code.

But in this case, if someone asks for ami-foo, and ami-foo doesn't
exist, that seems like it should throw an actual error code.